WIP - issues with the SCI/LIN on here, not sure what's going on.  LIN framing is being used on SCI so not sure what the fix is.
Added the GPS in - not really working

Added in the time module, nothing included yet

+++ source/sys_pcr.c
@@ -1,10 +1,10 @@
 /** @file sys_pcr.c
 *   @brief PCR Driver Implementation File
-*   @date 25.July.2013
-*   @version 03.06.00
+*   @date 9.Sep.2014
+*   @version 04.01.00
 *
 */
-/* (c) Texas Instruments 2009-2013, All rights reserved. */
+/* (c) Texas Instruments 2009-2014, All rights reserved. */
 
 #include "sys_pcr.h"
 
@@ -20,6 +20,9 @@
 *
 *   This function sets the protection for the selected frame.
 */
+/* SourceId : PCR_SourceId_001 */
+/* DesignId : PCR_DesignId_001 */
+/* Requirements : HL_SR41 */
 void peripheral_Frame_Protection_Set(peripheral_Frame_Select_t peripheral_Frame)
 {
 
@@ -62,6 +65,9 @@ void peripheral_Frame_Protection_Set(peripheral_Frame_Select_t peripheral_Frame)
 *
 *   This function clears the protection set for the selected frame.
 */
+/* SourceId : PCR_SourceId_002 */
+/* DesignId : PCR_DesignId_002 */
+/* Requirements : HL_SR42 */
 void peripheral_Frame_Protection_Clr(peripheral_Frame_Select_t peripheral_Frame)
 {
 
@@ -104,6 +110,9 @@ void peripheral_Frame_Protection_Clr(peripheral_Frame_Select_t peripheral_Frame)
 *
 *   This function will set the selected peripheral frame to powerdown.
 */
+/* SourceId : PCR_SourceId_003 */
+/* DesignId : PCR_DesignId_003 */
+/* Requirements : HL_SR43 */
 void peripheral_Frame_Powerdown_Set(peripheral_Frame_Select_t peripheral_Frame)
 {
 
@@ -146,6 +155,9 @@ void peripheral_Frame_Powerdown_Set(peripheral_Frame_Select_t peripheral_Frame)
 *
 *   This function will bring the selected peripheral frame out of powerdown.
 */
+/* SourceId : PCR_SourceId_004 */
+/* DesignId : PCR_DesignId_004 */
+/* Requirements : HL_SR44 */
 void peripheral_Frame_Powerdown_Clr(peripheral_Frame_Select_t peripheral_Frame)
 {
 
@@ -187,6 +199,9 @@ void peripheral_Frame_Powerdown_Clr(peripheral_Frame_Select_t peripheral_Frame)
 *
 *   This function sets the protection for the selected peripheral memory frame.
 */
+/* SourceId : PCR_SourceId_005 */
+/* DesignId : PCR_DesignId_017 */
+/* Requirements : HL_SR57 */
 void peripheral_Mem_Frame_Prot_Set(peripheral_MemoryFrame_CS_t peripheral_Memory_Frame_CS)
 {
 
@@ -199,11 +214,11 @@ void peripheral_Mem_Frame_Prot_Set(peripheral_MemoryFrame_CS_t peripheral_Memory
 
 	if (chip_select_grp >= 1U)
     {
-        pcrREG->PMPROTSET1 = (1U << (peripheral_Memory_Frame_CS & 0xFU));
+        pcrREG->PMPROTSET1 = (uint32)1U << (peripheral_Memory_Frame_CS & 0xFU);
     }
 	else
     {
-        pcrREG->PMPROTSET0 = (1U << peripheral_Memory_Frame_CS);
+        pcrREG->PMPROTSET0 = (uint32)1U << peripheral_Memory_Frame_CS;
     }
 	
 /* USER CODE BEGIN (15) */
@@ -219,6 +234,9 @@ void peripheral_Mem_Frame_Prot_Set(peripheral_MemoryFrame_CS_t peripheral_Memory
 *
 *   This function clears the protection set for the selected peripheral memory frame.
 */
+/* SourceId : PCR_SourceId_006 */
+/* DesignId : PCR_DesignId_018 */
+/* Requirements : HL_SR58 */
 void peripheral_Mem_Frame_Prot_Clr(peripheral_MemoryFrame_CS_t peripheral_Memory_Frame_CS)
 {
 
@@ -231,11 +249,11 @@ void peripheral_Mem_Frame_Prot_Clr(peripheral_MemoryFrame_CS_t peripheral_Memory
 
 	if (chip_select_grp >= 1U)
     {
-        pcrREG->PMPROTCLR1 = (1U << (peripheral_Memory_Frame_CS & 0xFU));
+        pcrREG->PMPROTCLR1 = (uint32)1U << (peripheral_Memory_Frame_CS & 0xFU);
     }
 	else
     {
-        pcrREG->PMPROTCLR0 = (1U << peripheral_Memory_Frame_CS);
+        pcrREG->PMPROTCLR0 = (uint32)1U << peripheral_Memory_Frame_CS;
     }
 
 /* USER CODE BEGIN (18) */
@@ -251,6 +269,9 @@ void peripheral_Mem_Frame_Prot_Clr(peripheral_MemoryFrame_CS_t peripheral_Memory
 *
 *   This function will set the selected peripheral memory frame to powerdown.
 */
+/* SourceId : PCR_SourceId_007 */
+/* DesignId : PCR_DesignId_019 */
+/* Requirements : HL_SR59 */
 void peripheral_Mem_Frame_Pwrdwn_Set(peripheral_MemoryFrame_CS_t peripheral_Memory_Frame_CS)
 {
 
@@ -263,11 +284,11 @@ void peripheral_Mem_Frame_Pwrdwn_Set(peripheral_MemoryFrame_CS_t peripheral_Memo
 
 	if (chip_select_grp >= 1U)
     {
-        pcrREG->PCSPWRDWNSET0 = (1U << (peripheral_Memory_Frame_CS & 0xFU));
+        pcrREG->PCSPWRDWNSET0 = (uint32)1U << (peripheral_Memory_Frame_CS & 0xFU);
     }
 	else
     {
-        pcrREG->PCSPWRDWNSET1 = (1U << peripheral_Memory_Frame_CS);
+        pcrREG->PCSPWRDWNSET1 = (uint32)1U << peripheral_Memory_Frame_CS;
     }
 	
 /* USER CODE BEGIN (21) */
@@ -283,6 +304,9 @@ void peripheral_Mem_Frame_Pwrdwn_Set(peripheral_MemoryFrame_CS_t peripheral_Memo
 *
 *   This function will bring the selected peripheral memory frame out of powerdown.
 */
+/* SourceId : PCR_SourceId_008 */
+/* DesignId : PCR_DesignId_020 */
+/* Requirements : HL_SR60 */
 void peripheral_Mem_Frame_Pwrdwn_Clr (peripheral_MemoryFrame_CS_t peripheral_Memory_Frame_CS)
 {
 
@@ -295,11 +319,11 @@ void peripheral_Mem_Frame_Pwrdwn_Clr (peripheral_MemoryFrame_CS_t peripheral_Mem
 
 	if (chip_select_grp >= 1U)
     {
-        pcrREG->PCSPWRDWNCLR0 = (1U << (peripheral_Memory_Frame_CS & 0xFU));
+        pcrREG->PCSPWRDWNCLR0 = (uint32)1U << (peripheral_Memory_Frame_CS & 0xFU);
     }
 	else
     {
-        pcrREG->PCSPWRDWNCLR1 = (1U << peripheral_Memory_Frame_CS);
+        pcrREG->PCSPWRDWNCLR1 = (uint32)1U << peripheral_Memory_Frame_CS;
     }
 	
 /* USER CODE BEGIN (24) */
@@ -315,6 +339,9 @@ void peripheral_Mem_Frame_Pwrdwn_Clr (peripheral_MemoryFrame_CS_t peripheral_Mem
 *
 *   This function sets the protection for all the selected frames.
 */
+/* SourceId : PCR_SourceId_009 */
+/* DesignId : PCR_DesignId_005 */
+/* Requirements : HL_SR45 */
 void peripheral_Protection_Set(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 {
 
@@ -339,6 +366,9 @@ void peripheral_Protection_Set(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 *
 *   This function clears the protection set for all the selected frame.
 */
+/* SourceId : PCR_SourceId_010 */
+/* DesignId : PCR_DesignId_006 */
+/* Requirements : HL_SR46 */
 void peripheral_Protection_Clr(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 {
 
@@ -363,6 +393,9 @@ void peripheral_Protection_Clr(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 *
 *   This function will set all the selected peripheral frame to powerdown.
 */
+/* SourceId : PCR_SourceId_011 */
+/* DesignId : PCR_DesignId_008 */
+/* Requirements : HL_SR48 */
 void peripheral_Powerdown_Set(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 {
 
@@ -387,6 +420,9 @@ void peripheral_Powerdown_Set(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 *
 *   This function will bring all the selected peripheral frame out of powerdown.
 */
+/* SourceId : PCR_SourceId_012 */
+/* DesignId : PCR_DesignId_009 */
+/* Requirements : HL_SR49 */
 void peripheral_Powerdown_Clr(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 {
 
@@ -411,6 +447,9 @@ void peripheral_Powerdown_Clr(peripheral_Quad_ChipSelect_t peripheral_Quad_CS)
 *
 *   This function sets the protection for all the selected peripheral memory frame.
 */
+/* SourceId : PCR_SourceId_013 */
+/* DesignId : PCR_DesignId_011 */
+/* Requirements : HL_SR51 */
 void peripheral_Memory_Protection_Set(peripheral_Memory_ChipSelect_t peripheral_Memory_CS)
 {
 
@@ -433,6 +472,9 @@ void peripheral_Memory_Protection_Set(peripheral_Memory_ChipSelect_t peripheral_
 *
 *   This function clears the protection set for all the selected peripheral memory frame.
 */
+/* SourceId : PCR_SourceId_014 */
+/* DesignId : PCR_DesignId_012 */
+/* Requirements : HL_SR52 */
 void peripheral_Memory_Protection_Clr(peripheral_Memory_ChipSelect_t peripheral_Memory_CS)
 {
 
@@ -455,6 +497,9 @@ void peripheral_Memory_Protection_Clr(peripheral_Memory_ChipSelect_t peripheral_
 *
 *   This function will set all the selected peripheral memory frame to powerdown.
 */
+/* SourceId : PCR_SourceId_015 */
+/* DesignId : PCR_DesignId_014 */
+/* Requirements : HL_SR54 */
 void peripheral_Memory_Powerdown_Set(peripheral_Memory_ChipSelect_t peripheral_Memory_CS)
 {
 
@@ -477,6 +522,9 @@ void peripheral_Memory_Powerdown_Set(peripheral_Memory_ChipSelect_t peripheral_M
 *
 *   This function will bring all the selected peripheral memory frame out of powerdown.
 */
+/* SourceId : PCR_SourceId_016 */
+/* DesignId : PCR_DesignId_015 */
+/* Requirements : HL_SR55 */
 void peripheral_Memory_Powerdown_Clr(peripheral_Memory_ChipSelect_t peripheral_Memory_CS)
 {
 
@@ -495,10 +543,13 @@ void peripheral_Memory_Powerdown_Clr(peripheral_Memory_ChipSelect_t peripheral_M
 
 /** @fn void peripheral_Powerdown_Status(peripheral_Quad_ChipSelect_t* peripheral_Quad_CS)
 *   @brief Get the powerdown status of the peripheral frames.
-*   @param[out] Peripheral frames power down status
+*   @param[out] peripheral_Quad_CS Peripheral frames power down status
 *
 *   This function gets the powerdown status of the peripheral frames.
 */
+/* SourceId : PCR_SourceId_017 */
+/* DesignId : PCR_DesignId_010 */
+/* Requirements : HL_SR50 */
 void peripheral_Powerdown_Status(peripheral_Quad_ChipSelect_t* peripheral_Quad_CS)
 {
 
@@ -520,10 +571,13 @@ void peripheral_Powerdown_Status(peripheral_Quad_ChipSelect_t* peripheral_Quad_C
 
 /** @fn void peripheral_Protection_Status(peripheral_Quad_ChipSelect_t* peripheral_Quad_CS )
 *   @brief Get the protection status of the peripheral frames
-*   @param[out] Peripheral frames protection status
+*   @param[out] peripheral_Quad_CS Peripheral frames protection status
 *
 *   This function gets the protection status of the peripheral frames.
 */
+/* SourceId : PCR_SourceId_018 */
+/* DesignId : PCR_DesignId_007 */
+/* Requirements : HL_SR47 */
 void peripheral_Protection_Status(peripheral_Quad_ChipSelect_t* peripheral_Quad_CS)
 {
 
@@ -545,10 +599,13 @@ void peripheral_Protection_Status(peripheral_Quad_ChipSelect_t* peripheral_Quad_
 
 /** @fn void peripheral_Memory_Protection_Status(peripheral_Memory_ChipSelect_t* peripheral_Memory_CS)
 *   @brief Get the protection set of all the peripheral Memory frame
-*   @param[out] Peripheral memory frames protection status
+*   @param[out] peripheral_Memory_CS Peripheral memory frames protection status
 *   
 *   This function gets the protection status of all the peripheral Memory frame.
 */
+/* SourceId : PCR_SourceId_019 */
+/* DesignId : PCR_DesignId_013 */
+/* Requirements : HL_SR53 */
 void peripheral_Memory_Protection_Status(peripheral_Memory_ChipSelect_t* peripheral_Memory_CS)
 {
 
@@ -569,10 +626,13 @@ void peripheral_Memory_Protection_Status(peripheral_Memory_ChipSelect_t* periphe
 
 /** @fn void peripheral_Memory_Powerdown_Status(peripheral_Memory_ChipSelect_t* peripheral_Memory_CS)
 *   @brief Get the powerdown status of all the peripheral Memory frame
-*   @param[out] Peripheral memory frames powerdown status
+*   @param[out] peripheral_Memory_CS Peripheral memory frames powerdown status
 *
 *   This function gets the powerdown status of all the peripheral Memory frame.
 */
+/* SourceId : PCR_SourceId_020 */
+/* DesignId : PCR_DesignId_016 */
+/* Requirements : HL_SR56 */
 void peripheral_Memory_Powerdown_Status(peripheral_Memory_ChipSelect_t* peripheral_Memory_CS)
 {
 
@@ -605,7 +665,9 @@ void peripheral_Memory_Powerdown_Status(peripheral_Memory_ChipSelect_t* peripher
 *   of the configuration registers to the struct pointed by config_reg
 *
 */
-
+/* SourceId : PCR_SourceId_021 */
+/* DesignId : PCR_DesignId_021 */
+/* Requirements : HL_SR61 */
 void pcrGetConfigValue(pcr_config_reg_t *config_reg, config_value_type_t type)
 {
 	if (type == InitialValue)
@@ -613,7 +675,7 @@ void pcrGetConfigValue(pcr_config_reg_t *config_reg, config_value_type_t type)
 	}
 	else
 	{
-	/*SAFETYMCUSW 134 S MR:12.2 <APPROVED> "Register read back support" */
+	/*SAFETYMCUSW 134 S MR:12.2 <APPROVED> "LDRA Tool issue" */
         config_reg->CONFIG_PMPROTSET0    = pcrREG->PMPROTSET0;
         config_reg->CONFIG_PMPROTSET1    = pcrREG->PMPROTSET1;
         config_reg->CONFIG_PPROTSET0     = pcrREG->PPROTSET0;
